requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
drf-standardized-errors==0.12.2
Markdown==3.4.1
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
